+++ ModLoader.Core/DefaultMemory.cs
@@ -7,102 +7,102 @@ public class DefaultMemory : IMemory
         throw new NotImplementedException();
     }
 
-    public static float ReadF32(uint address)
+    public static float ReadF32(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static double ReadF64(uint address)
+    public static double ReadF64(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static short ReadS16(uint address)
+    public static short ReadS16(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static int ReadS32(uint address)
+    public static int ReadS32(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static long ReadS64(uint address)
+    public static long ReadS64(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static sbyte ReadS8(uint address)
+    public static sbyte ReadS8(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static ushort ReadU16(uint address)
+    public static ushort ReadU16(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static uint ReadU32(uint address)
+    public static uint ReadU32(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static ulong ReadU64(uint address)
+    public static ulong ReadU64(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static byte ReadU8(uint address)
+    public static byte ReadU8(u64 address)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteF32(uint address, float value)
+    public static void WriteF32(u64 address, float value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteF64(uint address, double value)
+    public static void WriteF64(u64 address, double value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteS16(uint address, short value)
+    public static void WriteS16(u64 address, short value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteS32(uint address, int value)
+    public static void WriteS32(u64 address, int value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteS64(uint address, long value)
+    public static void WriteS64(u64 address, long value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteS8(uint address, sbyte value)
+    public static void WriteS8(u64 address, sbyte value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteU16(uint address, ushort value)
+    public static void WriteU16(u64 address, ushort value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteU32(uint address, uint value)
+    public static void WriteU32(u64 address, uint value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteU64(uint address, ulong value)
+    public static void WriteU64(u64 address, ulong value)
     {
         throw new NotImplementedException();
     }
 
-    public static void WriteU8(uint address, byte value)
+    public static void WriteU8(u64 address, byte value)
     {
         throw new NotImplementedException();
     }
